tests/wpt/tests/html/dom/render-blocking/script-inserted-stylesheet-link.html

<!doctype html>
<title>Script-inserted stylesheet links with "blocking=render" are render-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-render-blocking.js"></script>

<script>
const stylesheet = document.createElement('link');
stylesheet.rel = 'stylesheet';
stylesheet.href = 'support/target-red.css?pipe=trickle(d1)';
stylesheet.blocking = 'render';
document.head.appendChild(stylesheet);
</script>

<div class="target">
  This should be red
</div>

<script>
test_render_blocking(
    stylesheet,
    () => {
      let color = getComputedStyle(document.querySelector('.target')).color;
      assert_equals(color, 'rgb(255, 0, 0)');
    },
    'Render-blocking stylesheet is applied');
</script>
