LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/script-inserted-stylesheet-link-expected.txt

This should be red

PASS Rendering is blocked before render-blocking resources are loaded
PASS Render-blocking stylesheet is applied
